modyfikacje game engibe

+++ GameEngine/src/com/pong/gameengine/GameInterface.java
@@ -2,11 +2,11 @@ package com.pong.gameengine;
 
 public interface GameInterface {
 
-    void ballPosition(int x, int y);
+    void ballPosition(double x, double y);
 
-    void positionPlayer1(int y);
+    void positionPlayer1(double y);
 
-    void positionPlayer2(int y);
+    void positionPlayer2(double y);
 
     void gameScore(int player1Score, int player2Score);
